@@ -35,19 +35,18 @@
   },
   "dependencies": {
     "@agm/core": "^1.0.0-beta.5",
-    "@angular-devkit/build-angular": "^0.13.0",
-    "@angular-devkit/build-ng-packagr": "^0.803.21",
-    "@angular/animations": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular-devkit/build-ng-packagr": "~0.803.29",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@google/markerclusterer": "^2.0.0",
     "@nebular/auth": "3.0.0",
     "@nebular/bootstrap": "3.0.0",
@@ -76,7 +75,7 @@
     "localforage": "^1.8.1",
     "material-design-icons": "^3.0.1",
     "moment": "^2.24.0",
-    "ng-packagr": "^4.7.1",
+    "ng-packagr": "^5.4.0",
     "ng-pick-datetime": "^7.0.0",
     "ng2-completer": "^2.0.8",
     "ng2-pdf-viewer": "^5.3.4",
@@ -87,7 +86,7 @@
     "normalize.css": "6.0.0",
     "pace-js": "1.0.2",
     "roboto-fontface": "0.8.0",
-    "rxjs": "6.3.0",
+    "rxjs": "6.6.3",
     "rxjs-compat": "^6.0.0",
     "sass-loader": "^9.0.2",
     "stream": "0.0.2",
@@ -95,21 +94,21 @@
     "topojson-client": "^3.1.0",
     "tslib": "^1.13.0",
     "typeface-exo": "0.0.22",
-    "web-animations-js": "2.2.5",
+    "web-animations-js": "^2.3.2",
     "xlsx": "^0.14.3",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/core": "^10.0.4",
-    "@angular/cli": "^7.3.8",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "7.0.0",
+    "@angular/cli": "^8.3.29",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "8.2.14",
     "@compodoc/compodoc": "^1.1.10",
     "@types/d3": "^5.7.2",
     "@types/jasmine": "2.5.54",
     "@types/jasminewd2": "2.0.3",
     "@types/node": "6.0.90",
-    "codelyzer": "^4.5.0",
+    "codelyzer": "^5.0.1",
     "conventional-changelog-cli": "1.3.4",
     "husky": "0.13.3",
     "jasmine-core": "2.6.4",
@@ -126,6 +125,6 @@
     "stylelint": "^9.10.1",
     "ts-node": "3.2.2",
     "tslint": "5.7.0",
-    "typescript": "3.1.3"
+    "typescript": "3.5.3"
   }
 }
