@@ -15,82 +15,82 @@ import { CheckloginandredirectComponent } from './auth/checkloginandredirect/che
 const routes: Routes = [
   {
     path: 'pages',
-    loadChildren: 'app/pages/pages.module#PagesModule',
+    loadChildren: () => import('app/pages/pages.module').then(m => m.PagesModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'load-intelligence',
-    loadChildren: 'app/load-intelligence/load-intelligence.module#LoadIntelligenceModule',
+    loadChildren: () => import('app/load-intelligence/load-intelligence.module').then(m => m.LoadIntelligenceModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'intelligence',
-    loadChildren: 'app/intelligence/intelligence.module#IntelligenceModule',
+    loadChildren: () => import('app/intelligence/intelligence.module').then(m => m.IntelligenceModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'accounts',
-    loadChildren: 'app/accounts/accounts.module#AccountsModule',
+    loadChildren: () => import('app/accounts/accounts.module').then(m => m.AccountsModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'partner',
-    loadChildren: 'app/partner/partner.module#PartnerModule',
+    loadChildren: () => import('app/partner/partner.module').then(m => m.PartnerModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'admin',
-    loadChildren: 'app/admin/admin.module#AdminModule',
+    loadChildren: () => import('app/admin/admin.module').then(m => m.AdminModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'tyres',
-    loadChildren: 'app/tyres/tyres.module#TyresModule',
+    loadChildren: () => import('app/tyres/tyres.module').then(m => m.TyresModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'battery',
-    loadChildren: 'app/battery/battery.module#BatteryModule',
+    loadChildren: () => import('app/battery/battery.module').then(m => m.BatteryModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'ware-house',
-    loadChildren: 'app/ware-house/ware-house.module#WareHouseModule',
+    loadChildren: () => import('app/ware-house/ware-house.module').then(m => m.WareHouseModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'documents',
-    loadChildren: 'app/documents/documents.module#DocumentsModule',
+    loadChildren: () => import('app/documents/documents.module').then(m => m.DocumentsModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'lorry-receipt',
-    loadChildren: 'app/lorry-receipt/lorry-receipt.module#LorryReceiptModule',
+    loadChildren: () => import('app/lorry-receipt/lorry-receipt.module').then(m => m.LorryReceiptModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'driver',
-    loadChildren: 'app/driver/driver.module#DriverModule',
+    loadChildren: () => import('app/driver/driver.module').then(m => m.DriverModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'vehicle-maintenance',
-    loadChildren: 'app/vehicle-maintenance/vehicle-maintenance.module#VehicleMaintenanceModule',
+    loadChildren: () => import('app/vehicle-maintenance/vehicle-maintenance.module').then(m => m.VehicleMaintenanceModule),
     canActivate: [AuthGuard],
   },
   {
     path: 'walle8',
-    loadChildren: 'app/walle8/walle8.module#Walle8Module',
+    loadChildren: () => import('app/walle8/walle8.module').then(m => m.Walle8Module),
     canActivate: [AuthGuard],
   },
   {
     path:'challan',
-    loadChildren:'app/challan/challan.module#ChallanModule',
+    loadChildren:() => import('app/challan/challan.module').then(m => m.ChallanModule),
     canActivate:[AuthGuard]
   },
   {
     path: 'bid-system',
-    loadChildren: 'app/bid-system/bid-system.module#BidSystemModule',
+    loadChildren: () => import('app/bid-system/bid-system.module').then(m => m.BidSystemModule),
     canActivate: [AuthGuard],
   },
 
